refactor(infra): split infra into files
