+++ api/.golangci.yaml
